Code clean-up: separate cache_hacks into C and H files
